Add ka id to classe

+++ src/components/pages/classes.js
@@ -2,12 +2,13 @@ import React, { useState } from 'react'
 import { connect } from 'react-redux'
 import { withRouter } from 'react-router-dom'
 
+import ConfirmDialog from 'components/shared/dialogs/confirm-dialog'
 import MaterialTable from 'material-table'
 import Page from 'components/shared/page'
-import TextFieldDialog from 'components/shared/dialogs/text-field-dialog'
+import { TextField, makeStyles } from '@material-ui/core'
 
 import { createClass, deleteClass, fetchClasses, updateClass } from 'services/classes-service'
-import { flashError } from 'components/global-flash'
+import { flashError, flashSuccess } from 'components/global-flash'
 
 const mapDispatchToProps = dispatch => ({
   actions: {
@@ -22,10 +23,19 @@ const mapStateToProps = state => ({
   classes: Object.values(state.Classes),
 })
 
+const useStyles = makeStyles(theme => ({
+  newClassNameField: {
+    marginBottom: theme.spacing(1),
+  },
+}))
+
 function Classes(props) {
+  const classes = useStyles()
   const { actions } = props
   const [createDialogOpen, setCreateDialogOpen] = useState(false)
   const [createDialogLoading, setCreateDialogLoading] = useState(false)
+  const [newClassName, setNewClassName] = useState('')
+  const [newClassKAID, setNewClassKAID] = useState('')
   const [hasFetchedData, setHasFetchedData] = useState(false)
 
   if (!hasFetchedData) {
@@ -36,15 +46,18 @@ function Classes(props) {
 
   return (
     <Page>
-      <TextFieldDialog
+      <ConfirmDialog
         loading={ createDialogLoading }
         onClose={ () => setCreateDialogOpen(false) }
-        onConfirm={ name => {
+        onConfirm={ () => {
           setCreateDialogLoading(true)
-          actions.createClass({ name })
+          actions.createClass({ name: newClassName, ka_id: newClassKAID })
             .then(() => {
               setCreateDialogLoading(false)
               setCreateDialogOpen(false)
+              setNewClassName('')
+              setNewClassKAID('')
+              flashSuccess('Class created')
             })
             .catch(err => {
               flashError(err.message)
@@ -52,12 +65,24 @@ function Classes(props) {
             })
         } }
         open={ createDialogOpen }
-        textFieldProps={ {
-          autoCapitalize: 'words',
-          label: 'Name',
-        } }
         title='Create New Class'
-      />
+      >
+        <TextField 
+          autoCapitalize='words'
+          className={ classes.newClassNameField }
+          fullWidth
+          label='Name'
+          onChange={ e => setNewClassName(e.target.value) }
+          value={ newClassName }
+        />
+        <TextField 
+          autoCapitalize='none'
+          fullWidth
+          label='Khan Academy ID'
+          onChange={ e => setNewClassKAID(e.target.value) }
+          value={ newClassKAID }
+        />
+      </ConfirmDialog>
       <MaterialTable
         actions={ [
           {
@@ -69,15 +94,21 @@ function Classes(props) {
         columns={ [
           { title: 'ID', field: 'id', defaultSort: 'asc', editable: false },
           { title: 'Name', field: 'name' },
+          { title: 'Khan Academy ID', field: 'ka_id' },
         ] }
         data={ props.classes }
         editable={ {
           onRowDelete: async rowData => {
-            actions.deleteClass(rowData.id)
-              .catch(err => flashError(err.message))
+            if (confirm('ARE YOU SURE?')) {
+              actions.deleteClass(rowData.id)
+                .then(() => flashSuccess('Class deleted'))
+                .catch(err => flashError(err.message))
+            }
           },
           onRowUpdate: async rowData => {
-            actions.updateClass(rowData.id, rowData.name)
+            const { name, ka_id } = rowData
+            actions.updateClass(rowData.id, { name, ka_id })
+              .then(() => flashSuccess('Class updated'))
               .catch(err => flashError(err.message))
           },
         } }
@@ -85,7 +116,7 @@ function Classes(props) {
         options={ {
           search: false,
           paging: false,
-          actionsColumnIndex: 2,
+          actionsColumnIndex: 3,
         } }
         title='Class Sections'
       />
